Refactor game data world queries into module
